potts -> poly
